delete /root/.spot-status if created during pinstall.sh

+++ woof-code/rootfs-petbuilds/pa-applet/pinstall.sh
@@ -1,3 +1,3 @@
-mv -f root/.spot-status root/.spot-status.orig
+[ -f root/.spot-status ] && mv -f root/.spot-status root/.spot-status.orig
 chroot . /usr/sbin/setup-spot pa-applet=true
-mv -f root/.spot-status.orig root/.spot-status
+[ -f root/.spot-status.orig ] && mv -f root/.spot-status.orig root/.spot-status || rm -f root/.spot-status
